Add headers and basic stack navivgation
